pinctrl: msm: Remove explicit barriers from mmio ops where unneeded
For the vast majority of mmio operations in this driver, explicit memory
barriers aren't needed either because a data dependency between a read
and write already exists, or because of the presence of the spin locks
which execute a full memory barrier.

Removing all the unneeded explicit barriers considerably reduces
overhead for pinctrl operations, which in turn benefits things like i2c.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
Change-Id: I5817956a33749dca0ed1cdb6a8c071e0140f2d54
Signed-off-by: kyvangka1610 <kyvangka2002@gmail.com>
Signed-off-by: HafizZiq <hafizuddinismail552@gmail.com>