sched/tune: Only force UX tasks to big cores
If we force top-app's boost value to be >0 we get excellent
jitter values due to top-app being rendered on big cores.

However along with the actual top app a load of crap ends up on
big cluster too. We can avoid this by checking if adj_score is 0,
which means that this app is actually the one that's currently on
the screen (or it's one of its child processes).

[ghostrider-reborn]
- no need to compare with original stune boost value, since we use
  stune boosting only for top-app atm

Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
[HafizZiq]
- re-apply f73d691f8c8ea2dede1b6490420ac10ead1657d8
Signed-off-by: HafizZiq <hafizuddinismail552@gmail.com>